RPC: GET_LIMIT, SET_LIMIT Updates to new RPC interface.
Removes distinct `limit`, `limit_up`, `limit_down` command and makes
them all go through `limit` by allowing limit to take the down and up
parameters.

    limit_up 123
    limit_down 456

becomes

    limit 0 123
    limit 456 0

or to set both at once:

    limit 456 123

Also improves documentation, and allows "default" to be specified for
either limit to return the limit to its default.  (-1 is also accepted
for this, but is a pain because on the command line it produces an
invalid option error unless you also prefix the command with `--` to
stop further arg parsing).

Also makes the command actually print an error (instead of nothing at
all) when given invalid arguments.  Whoa such innovation!